+++ packages/cloud-cognitive/src/components/SidePanel/SidePanel.js
@@ -46,6 +46,7 @@ export let SidePanel = React.forwardRef(
       navigationBackIconDescription,
       onNavigationBack,
       onRequestClose,
+      onUnmount,
       open,
       pageContentSelector,
       placement,
@@ -290,7 +291,10 @@ export let SidePanel = React.forwardRef(
 
     // initialize the side panel to close
     const onAnimationEnd = () => {
-      if (!open) setRender(false);
+      if (!open) {
+        onUnmount && onUnmount();
+        setRender(false);
+      }
       sidePanelRef.current.style.overflow = 'auto';
       sidePanelRef.current.style.overflowX = 'hidden';
       setAnimationComplete(true);
@@ -645,6 +649,12 @@ SidePanel.propTypes = {
    * This handler closes the modal, e.g. changing `open` prop.
    */
   onRequestClose: PropTypes.func,
+
+  /**
+   * Optional function called when the side panel exit animation is complete.
+   * This handler can be used for any state cleanup needed before the panel is removed from the DOM.
+   */
+  onUnmount: PropTypes.func,
 
   /**
    * Determines whether the side panel should render or not