+++ packages/cloud-cognitive/src/components/SidePanel/SidePanel.test.js
@@ -22,6 +22,7 @@ const sizes = ['xs', 'sm', 'md', 'lg', 'max'];
 const dataTestId = uuidv4();
 
 const onRequestCloseFn = jest.fn();
+const onUnmountFn = jest.fn();
 const renderSidePanel = ({ ...rest }, children = <p>test</p>) =>
   render(
     <SidePanel
@@ -53,7 +54,8 @@ const SlideIn = ({
       onRequestClose={onRequestCloseFn}
       slideIn
       pageContentSelector="#side-panel-test-page-content"
-      placement={placement}>
+      placement={placement}
+      onUnmount={onUnmountFn}>
       Content
     </SidePanel>
     <div id="side-panel-test-page-content" />
@@ -154,6 +156,7 @@ describe('SidePanel', () => {
     fireEvent.animationEnd(outerElement);
     const updatedStyles = getComputedStyle(pageContent);
     expect(updatedStyles.marginRight).toBe('0px');
+    expect(onUnmountFn).toHaveBeenCalled();
   });
 
   it('should render a right slide in panel version', async () => {